Serialization, Warped Magi Abilities, And Code Cleanup
Implimented a framework for basic Serialization system.

At the moment, it simply Serializer and Deserializer are adapters for BinaryWriter / BinaryReader to my interface.

This allows me to simply pass an object that can write / read information without having to constrict it to the BinaryReader/Writer Classes.

Added Abilities for Warped Magi
-> Infusion
-> Magical Shield
->Rift Walk

Made Healthable and Magicable Serializable

Made all IModules serializable and deserializable, considering undoing this (so modules that don't pass information at all can be skipped entirely), although at the moment, if nothing is overriden, nothing is serialized anyways.

Moved some code around in Trigger; isolated the code which finds the 'root' gameobject.