Fixed Namespace naming & Finished up key aspects of NetworkServer/Client
Time to test it.

Also; I like event based (we still 'poll' it, but now we only poll in one location) but I need an immutable stream wrapper or something; A PROM stream sounds like an apropriate name.

Im not a networking expert though; and im sure as I develop further; ill find out how god awful this solution is.