Client closing immediately?
It seems the TCP client is closing immediately after being accepted? I cant figure this out right now so ill just leave this here.

Also, yes, this is literally a bad chat app right now. ITs a proof of concept for chatting.

Also, every Moba has all-chat, its how you communicate that you're an A-hole to the rest of your team. A 'vital' aspect of an MOBA.