It Works But I Broke The Demo
I had the demo working fine after porting from the simple Chat to advance chat; using Raw Messages only, but then I decided to show off fancy shmancy user defined events; but it seems to be failing to do that, AND I FORGOT TO COMMIT THE WORKING DEMO. Such is life.

Also, reading messages is just a pain; having to Reseek, Create a reader/writer etc is alot of work; perhaps I need to create an extension method to CreateSerializer/Deserializer (wrapped in a using) and create a MessageData base class which will serialize/deserialize automatically

One thing that sucks about my design pattern is that Codes can mean different things on Client/Server; so it might be best to have a Call/Response Code paring, this isn't important in the grand scheme of things, I just noticed it.