Moving Things Around
Finally Moved Ability to modules; i felt that Abilitiable belonged there but Ability doesnt; but this is consistent with the rest of the modules; so its good enough.

The 'Sub' Ability parts have been moved to subinfo; they aren't required for an ability, just allow us to describe it better.

It's how I plan on replacing Ability View.

Also surpressed 0649 warning on RingRender

DurationTimer can now set the timer; which will allow me to reuse it for cooldowns.