Why Do I Do This To Myself
= Removed Common Ability Info, replaced it with Module Cache & AbilityHelper
This took longer than i thought; and just goes to show how bad my ability system is; Its versatile, but not reusable.

Furthermore; the ability target pattern is useless for my intended purpse; as no cast information is sent (only targets); when an ability is 'reflected' when gathering targets; allies will still be ignored and enemies will get hurt again! On top of this; spells will never affect the caster!

On the other side; spell UI is going well.