Cool 'Atlas' are batched + some quick thoughts
In hindsight id be surprised if a sprite atlas wasnt batched, but its still cool to find that out firsthand.

I'd love to keep working on this. (As of my test, cpu is 11ms and rendering is 8ms)

Faster than minecrafts tickspeed (50ms) but again; this is with barebones art and stuff.

Without understanding the frame debugger too well; i'd hazard a guess that the two big render blocks are UI and the Material.

I bet i could optimize (or just drop) the shadergraph toon shader for a standard one (or a barebones one since i seem to care so much).

As for UI; since everything is transparent, less sprites is more performant. Also batching sprites together saves a few draws.

I'll put this up, but won't continue working on it.
I'll probably just remerge master into this when I decide to tinker with optimizations again and work from there.