Forest  Guardian
Nurture/Neglect is a passivew which improves or hinders nearby units healing.

Like overheat, BlackRose is an ability that is a toggleable which alters other abilities behaviour.

Also togglable got an event to notify the new state of the toggle. For claritty, the before and after are sent to avoid confusion of what the event bool means.