Prototype for a vision system
The core idea is that a level bakes a heightmap of the visible area.

Then, using a brenshem algorithm, we can traverse the 'pixels' of the height map for LOS. Essentially; raycasting a grid.

While I was hoping I could get away with only raycasting to the ring of the circle; this leads to some cells not being raycasted.

An okay solution; which I didn't impliment, would be to perform a ring raycheck, and 'touch' all checked cells.
Then iterate over all cells in the vision, and raycast to all untouched cells.

Another potential solution would be to raycast to a thicker ring. Or include an additional cast against the corners of cells.