[INDEV] Dropped old code
Vision Height Grid got replaced; so I dropped it.
Mapper got merged with the height grid info as well.

VisionWorldTest needs to be refactored using an asset, but shouldn't change too much.

I'm still trying to figure out how to handle team vision in a memory-friendly manner; I could pack them as 4 seperate maps of 8 channels.

This would allow me to seperate static, mixed, dynamic, and non-los vision into seperate 'texture spaces' making it cheaper to repaint (fewer objects to iterate over on a change).

That solution sounds nice, but the common case is going to be dynamic units moving around.

I'd consider doing this all on the GPU but because we need this information to be readable it's gotta all be CPU.

I'm sure somebody's already solved this (this dates back to Warcraft 3 and Age Of Empires 2 and Warhammer 40K after all). But for some reason I can only find information for FOW on roguelikes, which will similar, I feel have the advantage of being turnbased or minimizing the number of dynamic units requiring LOS.