Baked data & Complaints
Grid is alot of space for fine edges; when alot of terrain shares the same height. An octree would probably be better.

Rotation/Scale makes it hard to have 'vision range' be meaningful; a radius of 1 is actual an elipsoid if XZ aren't equal. A VisionUnit should be a defined, uniform unit; not arbitrary, as it currently is.

Height doesn't matter in this case, but they way I check it is pretty bad; a raycast and converting the hitpoint to a height value would probably be better.

Lastly; like how i handled voxels; if i chunk the vision grid; I don't have to rebuild the entire grid if I alter a chunk; only actors which can see that chunk.